chore: Refactoring dual farms zap2 and removing confirmation modal
